crm/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crm"
version = "0.1.0"
description = "Course and data tooling for video-vocab"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.170.0",
    "openai>=1.107.0",
    "pydantic==2.11.1",
    "python-dotenv==1.1.0",
    "tqdm==4.67.1",
    "yt-dlp>=2025.8.22",
    "youtube-transcript-api==1.0.3",
]

[project.scripts]
crm = "crm.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/crm"]
